# replace puts with say, bump dependencies, performance improvement, remove deprecated test calls