--- a/bin/zat
+++ b/bin/zat
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 lib_dir = File.expand_path('../../lib', __FILE__)
-$LOAD_PATH << lib_dir unless $LOAD_PATH.include?(lib_dir)
+$LOAD_PATH.shift lib_dir unless $LOAD_PATH.include?(lib_dir)
 ENV['THOR_SHELL'] = 'Color'
 require 'zendesk_apps_tools'
 ZendeskAppsTools::Command.start