--- a/lib/zendesk_apps_tools/command_helpers.rb
+++ b/lib/zendesk_apps_tools/command_helpers.rb
@@ -4,7 +4,6 @@ require 'zendesk_apps_tools/api_connection'
 require 'zendesk_apps_tools/deploy'
 require 'zendesk_apps_tools/directory'
 require 'zendesk_apps_tools/package_helper'
-require 'zendesk_apps_tools/settings'
 require 'zendesk_apps_tools/translate'
 require 'zendesk_apps_tools/bump'