--- a/lib/zendesk_apps_tools/directory.rb
+++ b/lib/zendesk_apps_tools/directory.rb
@@ -11,10 +11,9 @@ module ZendeskAppsTools
     end
 
     def prompt_new_app_dir
-      prompt = "Enter a directory name to save the new app (will create the dir if it does not exist, default to current dir):\n"
-      opts = { valid_regex: /^(\w|\/|\\)*$/, allow_empty: true }
+      prompt = "Enter a directory name to save the new app (will create the dir if it does not exist):\n"
+      opts = { valid_regex: /^(.|\w|\/|\\)*$/, default: './' }
       while @app_dir = get_value_from_stdin(prompt, opts)
-        @app_dir = './' and break if @app_dir.empty?
         if !File.exist?(@app_dir)
           break
         elsif !File.directory?(@app_dir)