--- a/lib/zendesk_apps_tools/server.rb
+++ b/lib/zendesk_apps_tools/server.rb
@@ -5,16 +5,14 @@ require 'zendesk_apps_support/package'
 module ZendeskAppsTools
   class Server < Sinatra::Base
     set :protection, :except => :frame_options
-    last_mtime = Time.new(0)
     ZENDESK_DOMAINS_REGEX = /^http(?:s)?:\/\/[a-z0-9-]+\.(?:zendesk|zopim|zd-(?:dev|master|staging))\.com$/
 
     get '/app.js' do
       access_control_allow_origin
       content_type 'text/javascript'
 
-      if new_settings = settings.settings_helper.refresh!
-        settings.parameters = new_settings
-      end
+      new_settings = settings.settings_helper.refresh!
+      settings.parameters = new_settings if new_settings
 
       package = ZendeskAppsSupport::Package.new(settings.root, false)
       app_name = ENV.fetch('ZAT_APP_NAME', 'Local App')