--- a/lib/zendesk_apps_tools/settings.rb
+++ b/lib/zendesk_apps_tools/settings.rb
@@ -47,32 +47,15 @@ module ZendeskAppsTools
       return nil unless File.exist? filepath
 
       begin
-        @last_mtime = File.stat(filepath).mtime
-        settings_file = File.read(filepath)
-
-        if filepath =~ /\.json$/ || settings_file =~ /\A\s*{/
-          settings_data = JSON.load(settings_file)
-        else
-          settings_data = YAML.load(settings_file)
-        end
-
-        settings_data.each do |index, setting|
-          if setting.is_a?(Hash) || setting.is_a?(Array)
-            settings_data[index] = JSON.dump(setting)
-          end
-        end
+        settings_file = read_settings(filepath)
+        settings_data = parse_settings(filepath, settings_file)
       rescue => err
-        @cli.say_error "Failed to load #{filepath}"
-        @cli.say_error err.message
+        @cli.say_error "Failed to load #{filepath}\n#{err.message}"
         return nil
       end
 
-      parameters.inject({}) do |settings, param|
-        input = settings_data[param['name']]
-
-        if !input && param['default']
-          input = param['default']
-        end
+      parameters.each_with_object({}) do |settings, param|
+        input = settings_data[param['name']] || param['default']
 
         if !input && param['required']
           @cli.say_error "'#{param['name']}' is required but not specified in the config file.\n"
@@ -89,6 +72,26 @@ module ZendeskAppsTools
     end
 
     private
+
+    def read_settings(filepath)
+      @last_mtime = File.stat(filepath).mtime
+      File.read(filepath)
+    end
+
+    def parse_settings(filepath, contents)
+      settings_data =
+        if filepath =~ /\.json$/ || contents =~ /\A\s*{/
+          JSON.load(contents)
+        else
+          YAML.load(contents)
+        end
+      settings_data.each do |index, setting|
+        if setting.is_a?(Hash) || setting.is_a?(Array)
+          settings_data[index] = JSON.dump(setting)
+        end
+      end
+      settings_data
+    end
 
     def convert_to_boolean_for_checkbox(input)
       unless [TrueClass, FalseClass].include?(input.class)