--- a/spec/lib/zendesk_apps_tools/command_spec.rb
+++ b/spec/lib/zendesk_apps_tools/command_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 require 'command'
 
 describe ZendeskAppsTools::Command do
-  PREFIX = 'https://username:password@subdomain.zendesk.com'
+  PREFIX = 'https://subdomain.zendesk.com'
 
   before do
     @command = ZendeskAppsTools::Command.new
@@ -25,6 +25,7 @@ describe ZendeskAppsTools::Command do
         allow(Faraday::UploadIO).to receive(:new)
 
         stub_request(:post, PREFIX + '/api/v2/apps/uploads.json')
+          .with(headers: { 'Authorization' => 'Basic dXNlcm5hbWU6cGFzc3dvcmQ=' })
           .to_return(body: '{ "id": 123 }')
 
         expect(@command.upload('nah')).to eq(123)
@@ -37,6 +38,7 @@ describe ZendeskAppsTools::Command do
         expect(Faraday::UploadIO).to receive(:new).with('app.zip', 'application/zip').and_return(nil)
 
         stub_request(:post, PREFIX + '/api/v2/apps/uploads.json')
+          .with(headers: { 'Authorization' => 'Basic dXNlcm5hbWU6cGFzc3dvcmQ=' })
           .to_return(body: '{ "id": 123 }')
 
         expect(@command.upload('nah')).to eq(123)
@@ -52,7 +54,8 @@ describe ZendeskAppsTools::Command do
         expect(File).to receive(:read) { '{ "name": "abc" }' }
 
         stub_request(:post, PREFIX + '/api/v2/apps.json')
-          .with(body: JSON.generate(name: 'abc', upload_id: '123'))
+          .with(body: JSON.generate(name: 'abc', upload_id: '123'),
+                headers: { 'Authorization' => 'Basic dXNlcm5hbWU6cGFzc3dvcmQ=' })
 
         @command.create
       end
@@ -67,7 +70,8 @@ describe ZendeskAppsTools::Command do
         expect(@command).to receive(:get_value_from_stdin) { 'abc' }
 
         stub_request(:post, PREFIX + '/api/v2/apps.json')
-          .with(body: JSON.generate(name: 'abc', upload_id: '123'))
+          .with(body: JSON.generate(name: 'abc', upload_id: '123'),
+                headers: { 'Authorization' => 'Basic dXNlcm5hbWU6cGFzc3dvcmQ=' })
 
         @command.create
       end
@@ -82,6 +86,7 @@ describe ZendeskAppsTools::Command do
         expect(@command).to receive(:fetch_cache).with('app_id').and_return(456)
 
         stub_request(:put, PREFIX + '/api/v2/apps/456.json')
+          .with(headers: { 'Authorization' => 'Basic dXNlcm5hbWU6cGFzc3dvcmQ=' })
 
         @command.update
       end
@@ -101,6 +106,7 @@ describe ZendeskAppsTools::Command do
         }
 
         stub_request(:get, PREFIX + '/api/v2/apps.json')
+          .with(headers: { 'Authorization' => 'Basic dXNlcm5hbWU6cGFzc3dvcmQ=' })
           .to_return(body: JSON.generate(apps))
 
         expect(@command.send(:find_app_id)).to eq(125)