--- a/zendesk_apps_tools.gemspec
+++ b/zendesk_apps_tools.gemspec
@@ -27,6 +27,7 @@ Gem::Specification.new do |s|
   s.add_development_dependency 'rspec'
   s.add_development_dependency 'bump'
   s.add_development_dependency 'webmock'
+  s.add_development_dependency 'byebug'
 
   s.files        = Dir.glob('{bin,lib,app_template*,templates}/**/*') + %w(README.md LICENSE)
   s.test_files   = Dir.glob('features/**/*')