… Gemfile.lock
… spec/lib/zendesk_apps_tools/settings_spec.rb