@sandlerr 2016-09-01
I bumped cucumber, now it raises if we end in a non-existent directory due to `pwd`: https://github.com/cucumber/cucumber-ruby/blob/master/lib/cucumber/formatter/backtrace_filter.rb#L21