@sandlerr 2016-09-01
The increased complexity was for a performance boost, `zat` is currently super slow to just load let alone do any actual tasks. This is *very* noticeable when running `zat version` or an unknown command:

#### Master:

```
zat asdf  1.04s user 0.42s system 42% cpu 3.435 total
zat asdf  1.04s user 0.39s system 51% cpu 2.753 total
zat asdf  1.04s user 0.37s system 40% cpu 3.501 total
zat asdf  1.02s user 0.36s system 43% cpu 3.156 total
```

#### This branch:

```
zat asdf  0.42s user 0.19s system 44% cpu 1.393 total
zat asdf  0.42s user 0.17s system 48% cpu 1.223 total
zat asdf  0.44s user 0.18s system 64% cpu 0.971 total
```